+++ package.json
@@ -25,20 +25,20 @@
     "@radix-ui/react-tooltip": "^1.1.3",
     "@supabase/ssr": "^0.5.1",
     "@supabase/supabase-js": "^2.45.6",
-    "@vercel/analytics": "^1.3.1",
+    "@vercel/analytics": "^1.3.2",
     "@vercel/postgres": "^0.10.0",
-    "@vercel/speed-insights": "^1.0.13",
+    "@vercel/speed-insights": "^1.0.14",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
     "embla-carousel-react": "^8.3.0",
-    "framer-motion": "^11.11.9",
+    "framer-motion": "^11.11.10",
     "geist": "^1.3.1",
     "lucide-react": "^0.453.0",
-    "next": "15.0.0",
+    "next": "15.0.1",
     "next-themes": "^0.3.0",
     "react": "19.0.0-rc-65a56d0e-20241020",
-    "react-day-picker": "9.1.4",
+    "react-day-picker": "9.2.0",
     "react-dom": "19.0.0-rc-65a56d0e-20241020",
     "react-feather": "^2.0.10",
     "sharp": "^0.33.5",
@@ -51,13 +51,13 @@
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@types/node": "22.7.8",
+    "@types/node": "22.8.1",
     "@types/react": "npm:types-react@19.0.0-rc.1",
     "@types/react-dom": "npm:types-react-dom@19.0.0-rc.1",
     "autoprefixer": "^10.4.20",
     "encoding": "^0.1.13",
     "eslint": "^9.13.0",
-    "eslint-config-next": "15.0.0",
+    "eslint-config-next": "15.0.1",
     "postcss": "^8.4.47",
     "prettier": "^3.3.3"
   },
